+++ zenodot-tests/test/dd/kms/zenodot/tests/completionTests/ClassTest.java
@@ -25,7 +25,10 @@ public class ClassTest extends CompletionTest
 
 		String className = ClassUtils.getRegularClassName(TestClass.class.getName());
 		testBuilder
-			.configurator(test -> test.minimumAccessModifier(AccessModifier.PACKAGE_PRIVATE))
+			.configurator(test -> {
+				test.minimumFieldAccessModifier(AccessModifier.PACKAGE_PRIVATE);
+				test.minimumMethodAccessModifier(AccessModifier.PACKAGE_PRIVATE);
+			})
 			.addTest(className + ".",				"f", "l", "getDouble()", "getInt()", "InnerClass")
 			.addTest(className + ".I",				"InnerClass")
 			.addTest(className + ".InnerClass.",	"test")
@@ -71,7 +74,7 @@ public class ClassTest extends CompletionTest
 			.addTest("InternalClassStage2.",	"i");
 
 		testBuilder
-			.configurator(test -> test.minimumAccessModifier(AccessModifier.PUBLIC))
+			.configurator(test -> test.minimumFieldAccessModifier(AccessModifier.PUBLIC))
 			.addUnstableTest("Ma",		"Math")
 			.addUnstableTest("Math.p",	"pow(, )", "PI")
 			.addUnstableTest("Math.P",	"PI", "pow(, )");